==> src/02_data_exploration.Rmd <==
---
title: "Data Exploration"
author: "Claire Kelling and Xiaoran Sun"
date: "2018/3/1"
output: html_document
editor_options: 
  chunk_output_type: console
---

##This is to read in and explore the subsetted data
```{r}
#set wherever your data directory is
#setwd("~/Desktop/paper/SODA501") #Xiaoran
setwd("C:/Users/ckell/OneDrive/Penn State/2017-2018/01_Spring/SODA_501/SODA_501_project/data") #Claire
library(sp)
library(ggplot2)
library(ggmap)
library(dplyr)

load(file = "coll_town_tweets.Rdata")
#load(file = "user_locations_finalcleaned.Rdata")

length(unique(coll_town_user_tweets$user$id))

towns<-c("Ithaca, NY", "State College, PA", "Bloomington, IN", "Lawrence,
         KS", "Blacksburg, VA", "College Station, TX", "Columbia, MO",
         "Champaign, IL", "Ann Arbor, MI", "Gainesville, FL")
tweets_in_college_town<-coll_town_user_tweets[grep(paste(towns,collapse="|"),
                                              coll_town_user_tweets$place$full_name),]
```

Some details about the dataset:
* There are 346,749 tweets in our dataset.
* There are 2234 users that have at some point tweeted in college towns.
* There are 11,786 tweets that were tweeted in a college town.

Now, I would like to see where a given user is tweeting. I will just try to perform this analysis for one user, to start.
```{r}
first_user_id <- tweets_in_college_town[2,]
first_user_id2 <- first_user_id$user$id

first_user_tweets <- coll_town_user_tweets[which(coll_town_user_tweets$user$id == first_user_id2),]
#he has 20 tweets
unique(first_user_tweets$place$name)

places_first_user <- unique(first_user_tweets$place) # he went to 3 places
coords_first_user <- as.data.frame(first_user_tweets$coordinates$coordinates) # coordinates
coords_first_user <- as.data.frame(t(as.matrix(coords_first_user)))
colnames(coords_first_user) <- c("Long", "Lat")
rownames(coords_first_user) <- c()

#converting to spatial points, may use this later
## coordinates(coords_first_user) <- cbind(coords_first_user$Long, coords_first_user$Lat)
## proj4string(coords_first_user) <- CRS("+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
#PA_Map <- qmap('pennsylvania', legend = 'topleft')

us <- map_data("state")

arr <- USArrests %>% 
  add_rownames("region") %>% 
  mutate(region=tolower(region))

gg <- ggplot()
gg <- gg + #geom_map(data=us, map=us,
                    #aes(x=long, y=lat, map_id=region),
                    #fill="#ffffff", color="#ffffff", size=0.15)+ 
  geom_point(aes(x = Long, y = Lat), data = coords_first_user)+
  geom_path(data=coords_first_user, aes(x=Long, y=Lat), color="black", size=1)
#cannot see when including map

```

Now, I would like to implement this for a different user.
```{r}
first_user_id <- tweets_in_college_town[4,] #4 is a college student from Nova, 12 is pretty close (not a student)
first_user_id2 <- first_user_id$user$id

first_user_tweets <- coll_town_user_tweets[which(coll_town_user_tweets$user$id == first_user_id2),]

#What is this user tweeting? 
tweet_text <- unique(first_user_tweets$text)
tweet_location <- first_user_tweets$place$name

unique(first_user_tweets$place$name)

places_first_user <- unique(first_user_tweets$place) # he went to 3 places
coords_first_user <- as.data.frame(first_user_tweets$coordinates$coordinates) # coordinates
coords_first_user <- as.data.frame(t(as.matrix(coords_first_user)))
colnames(coords_first_user) <- c("Long", "Lat")
rownames(coords_first_user) <- c()


gg <- ggplot()
gg <- gg + geom_map(data=us, map=us,
                    aes(x=long, y=lat, map_id=region),
                    fill="#ffffff", color="#ffffff", size=0.15)+ 
  geom_point(aes(x = Long, y = Lat), data = coords_first_user)+
  geom_path(data=coords_first_user, aes(x=Long, y=Lat), color="blue", size=1)
gg

#This is almost certainly a college person going home to NOVA from Blacksburg!
```


Now, I would like to see where each user is tweeting. I will perform this analysis for the full dataset.
```{r}

full_df <- NULL
j <- 0
for(i in unique(coll_town_user_tweets$user$id)){
  #i<- coll_town_user_tweets$user$id[1] #testing
  user_tweets <- coll_town_user_tweets[which(coll_town_user_tweets$user$id == i),]
  num_places <- length(unique(user_tweets$place$name))
  
  #I'd like to find the college town that is affilliated with each user
  tweets_ctown<-user_tweets[grep(paste(towns,collapse="|"),
                                              user_tweets$place$full_name),]
  coll_town <- unique(tweets_ctown$place$name)
  
  if(num_places < 25){ #filter for the number of unique places in these days
    coords_user <- as.data.frame(user_tweets$coordinates$coordinates) # coordinates
    coords_user <- as.data.frame(t(as.matrix(coords_user)))
    colnames(coords_user) <- c("Long", "Lat")
    rownames(coords_user) <- c()
    group <- rep(i, nrow(coords_user))
    ctown <- rep(coll_town, nrow(coords_user))
    coords_user <- cbind(coords_user, group, ctown)
    
    if(nrow(coords_user) < 100 & nrow(coords_user)>1){ # this is the filter for the number of tweets
      full_df <- rbind(full_df, coords_user)
    }
  }
  j <- j+1
  print(j) #should get up to 2234
  print(nrow(full_df)) 
}
#full_df <- full_df[1:2000,]
full_df$group <-as.factor(full_df$group)
full_df$ctown <-as.factor(full_df$ctown)

gg_user <- ggplot()
gg_user <- gg_user + geom_map(data=us, map=us,
                    aes(x=long, y=lat, map_id=region),
                    fill="#ffffff", color="#ffffff", size=0.15)+ 
  #geom_point(aes(x = Long, y = Lat), data = full_df)+
  geom_path(data=full_df, aes(x=Long, y=Lat, group = group, color= group), size=1)+
  theme(legend.position="none")
gg_user

gg_ctown <- ggplot()
gg_ctown <- gg_ctown + geom_map(data=us, map=us,
                    aes(x=long, y=lat, map_id=region),
                    fill="#ffffff", color="#ffffff", size=0.15)+ 
  #geom_point(aes(x = Long, y = Lat), data = full_df)+
  geom_path(data=full_df, aes(x=Long, y=Lat, group = ctown, color= ctown), size=1, alpha = 0.4)+
  facet_wrap(~ ctown, ncol = 3) #+ theme(legend.position="none")
gg_ctown #save as 1000x750

unique(full_df$ctown)
```
