==> src/09_four_missing_towns.Rmd <==
---
title: "09_four_missing_towns"
author: "Xiaoran Sun and Claire Kelling"
date: "2018/4/16"
output: html_document
---

# week April 16 updates:
We have four college towns for which there were ZERO users/tweets. Therefore, we want to look carefully into what's happending with these towns.
They are:
Ithaca, NY
Bloomington, IN
Lawrence, KS
Columbia, MO

```{r}
library(ggmap)
library(XML)
library(geosphere)
library(geonames)
library(Imap)
library(googleway)
library(dplyr)
library(xtable)
```

One way is to filter through the raw tweets using the state name and see what's the closest to those cities.
```{r}
#1. MO (seems like the easiest)
MOplaces<-unique(rawdat[grepl("MO", rawdat$place$full_name),]$place$full_name)
save(MOplaces, file = "/storage/home/xbs5014/Downloads/MOplaces.Rdata")
```

```{r}
#then, find out the closest places to Columbia, MO
## Using Claire's code in other section
#key <- 
#key2 <- 
load("~/Desktop/paper/SODA501/project/MOplaces.Rdata")
lonlat3 <- NULL
for(i in 1:length(MOplaces)){
  addr<-MOplaces[i]
  lonlat2 <- google_geocode(address = addr, key = key)
  lonlat2 <- lonlat2$results$geometry$location
  print(i)
  #Sys.sleep(3)
  lonlat_full<-cbind(addr, lonlat2)
  lonlat3 <- rbind(lonlat3, lonlat_full)
}


#dmat3 <- round(GeoDistanceInMetresMatrix(df.MOcities) / 1000) 
##This Claire's code doesn' work well for me, so I have to use distm(), which returns results in meters
lonlatColumbia<-google_geocode(address = "Columbia, MO", key = key2)
lonlatColumbia <- lonlatColumbia$results$geometry$location

lonlat3$DistToTown<-rep(NA, length(lonlat3$addr))
for(i in 1:length(lonlat3$addr)){
  dist_temp<-round(distm(c(lonlatColumbia$lng, lonlatColumbia$lat), c(lonlat3[i,]$lng, lonlat3[i,]$lat), fun = distHaversine)/1000, 2)
  lonlat3[i,]$DistToTown<-dist_temp
}

#sorting the output in ascending order
lonlat4<-lonlat3[order(lonlat3$DistToTown),]

#Found the closest: Ashland, MO, 20.83km away from Columbia, MO; the second closest is Centralia, MO, 33.41km away

MOtop5<-lonlat4[1:5, c(1,4)]
colnames(MOtop5)<-c("Town/City Name", "Distance to Columbia, MO")
xtable(MOtop5)

```

```{r}
#2. KS
KSplaces<-unique(rawdat[grepl("KS", rawdat$place$full_name),]$place$full_name)
save(KSplaces, file = "/storage/home/xbs5014/Downloads/KSplaces.Rdata")
```

```{r}
#then, find out the closest places to Lawrence, KS
## Using Claire's code in other section
#key <- 
#key2 <- 
load("~/Desktop/paper/SODA501/project/KSplaces.Rdata")
lonlat3 <- NULL
for(i in 1:length(KSplaces)){
  addr<-KSplaces[i]
  lonlat2 <- google_geocode(address = addr, key = key)
  lonlat2 <- lonlat2$results$geometry$location
  print(i)
  #Sys.sleep(3)
  lonlat_full<-cbind(addr, lonlat2)
  lonlat3 <- rbind(lonlat3, lonlat_full)
}


lonlatLawrence<-google_geocode(address = "Lawrence, KS", key = key2)
lonlatLawrence <- lonlatLawrence$results$geometry$location

lonlat3$DistToTown<-rep(NA, length(lonlat3$addr))
for(i in 1:length(lonlat3$addr)){
  dist_temp<-round(distm(c(lonlatLawrence$lng, lonlatLawrence$lat), c(lonlat3[i,]$lng, lonlat3[i,]$lat), fun = distHaversine)/1000, 2)
  lonlat3[i,]$DistToTown<-dist_temp
}

#sorting the output in ascending order
lonlat4<-lonlat3[order(lonlat3$DistToTown),]

#Found the closest: Eudora, KS, 12.24km away from Lawrence, KS; the second closest is Tonganoxie, KS, 19.97km away

KStop5<-lonlat4[1:5, c(1,4)]
colnames(KStop5)<-c("Town/City Name", "Distance to Lawrence, KS")
xtable(KStop5)

```

```{r}
#then, find out the closest places to Lawrence, KS
## Using Claire's code in other section
#key <- 
#key2 <- 
load("~/Desktop/paper/SODA501/project/KSplaces.Rdata")
lonlat3 <- NULL
for(i in 1:length(KSplaces)){
  addr<-KSplaces[i]
  lonlat2 <- google_geocode(address = addr, key = key)
  lonlat2 <- lonlat2$results$geometry$location
  print(i)
  #Sys.sleep(3)
  lonlat_full<-cbind(addr, lonlat2)
  lonlat3 <- rbind(lonlat3, lonlat_full)
}


lonlatLawrence<-google_geocode(address = "Lawrence, KS", key = key2)
lonlatLawrence <- lonlatLawrence$results$geometry$location

lonlat3$DistToTown<-rep(NA, length(lonlat3$addr))
for(i in 1:length(lonlat3$addr)){
  dist_temp<-round(distm(c(lonlatLawrence$lng, lonlatLawrence$lat), c(lonlat3[i,]$lng, lonlat3[i,]$lat), fun = distHaversine)/1000, 2)
  lonlat3[i,]$DistToTown<-dist_temp
}

#sorting the output in ascending order
lonlat4<-lonlat3[order(lonlat3$DistToTown),]

#Found the closest: Eudora, KS, 12.24km away from Lawrence, KS; the second closest is Tonganoxie, KS, 19.97km away

KStop5<-lonlat4[1:5, c(1,4)]
colnames(KStop5)<-c("Town/City Name", "Distance to Lawrence, KS")
xtable(KStop5)

```

```{r}
#then, find out the closest places to Bloomington, IN
## Using Claire's code in other section
#key <- 
#key2 <- 
load("~/Desktop/paper/SODA501/project/INplaces.Rdata")
lonlat3 <- NULL
for(i in 1:length(INplaces)){
  addr<-INplaces[i]
  lonlat2 <- google_geocode(address = addr, key = key)
  lonlat2 <- lonlat2$results$geometry$location
  print(i)
  #Sys.sleep(3)
  lonlat_full<-cbind(addr, lonlat2)
  lonlat3 <- rbind(lonlat3, lonlat_full)
}


lonlatBloomington<-google_geocode(address = "Bloomington, IN", key = key2)
lonlatBloomington <- lonlatBloomington$results$geometry$location

lonlat3$DistToTown<-rep(NA, length(lonlat3$addr))
for(i in 1:length(lonlat3$addr)){
  dist_temp<-round(distm(c(lonlatBloomington$lng, lonlatBloomington$lat), c(lonlat3[i,]$lng, lonlat3[i,]$lat), fun = distHaversine)/1000, 2)
  lonlat3[i,]$DistToTown<-dist_temp
}

#sorting the output in ascending order
lonlat4<-lonlat3[order(lonlat3$DistToTown),]

#Found the closest: Ellettsville, IN, 11.43km away from Bloomington, IN; the second closest is Oolitic, IN, 29.44km away

INtop5<-lonlat4[1:5, c(1,4)]
colnames(INtop5)<-c("Town/City Name", "Distance to Ellettsville, IN")
xtable(INtop5)

```

```{r}
#then, find out the closest places to Ithaca, NY
## Using Claire's code in other section
#key <- 
#key2 <- 
load("~/Desktop/paper/SODA501/project/NYplaces.Rdata")
lonlat3 <- NULL
for(i in 1:length(NYplaces)){
  addr<-NYplaces[i]
  lonlat2 <- google_geocode(address = addr, key = key)
  lonlat2 <- lonlat2$results$geometry$location
  print(i)
  #Sys.sleep(3)
  lonlat_full<-cbind(addr, lonlat2)
  lonlat3 <- rbind(lonlat3, lonlat_full)
}


lonlatIthaca<-google_geocode(address = "Ithaca, NY", key = key2)
lonlatIthaca <- lonlatIthaca$results$geometry$location

lonlat3$DistToTown<-rep(NA, length(lonlat3$addr))
for(i in 1:length(lonlat3$addr)){
  dist_temp<-round(distm(c(lonlatIthaca$lng, lonlatIthaca$lat), c(lonlat3[i,]$lng, lonlat3[i,]$lat), fun = distHaversine)/1000, 2)
  lonlat3[i,]$DistToTown<-dist_temp
}

#sorting the output in ascending order
lonlat4<-lonlat3[order(lonlat3$DistToTown),]

#Found the closest: Waverly, NY, 48.31km away from Ithaca, NY; the second closest is Vestal, NY, 54.36km away

NYtop5<-lonlat4[1:5, c(1,4)]
colnames(NYtop5)<-c("Town/City Name", "Distance to Ithaca, NY")
xtable(NYtop5)

```
